Fix incremental compilation bug
WARNING: Method definition (::Type{Judycon.QuickFind})(Any) in module
Judycon at /home/jukka/dev/Judycon/src/Judycon.jl:7 overwritten at
/home/jukka/dev/Judycon/src/Judycon.jl:11.
  ** incremental compilation may be fatally broken for this module **

Caused that there was function QuickFind(n), which is now
QuickFind(n::Int).

+++ src/Judycon.jl
@@ -7,7 +7,7 @@ struct QuickFind
     id :: Vector{Int}
 end
 
-function QuickFind(n)
+function QuickFind(n::Int)
     id = collect(1:n)
     return QuickFind(id)
 end
